+++ frontend/app/src/components/home/profile/SendCrypto.svelte
@@ -7,15 +7,12 @@
         type ResourceKey,
     } from "openchat-client";
     import {
-        ARBITRUM_NETWORK,
-        BASE_NETWORK,
         BTC_SYMBOL,
         CKBTC_SYMBOL,
         cryptoBalanceStore,
         cryptoLookup,
         currentUserIdStore,
         currentUserStore,
-        ETHEREUM_NETWORK,
         iconSize,
         ICP_SYMBOL,
         mobileWidth,
@@ -73,13 +70,13 @@
     let selectedNetwork = $state<string>();
     let isBtc = $derived(symbol === BTC_SYMBOL);
     let isBtcNetwork = $derived(selectedNetwork === BTC_SYMBOL);
-    let oneSecEnabled = $derived(tokenDetails.oneSecEnabled);
+    let oneSecEnabled = $derived(tokenDetails.oneSecEnabled && tokenDetails.evmContractAddresses.length > 0);
     let isOneSecNetwork = $derived(oneSecEnabled && selectedNetwork !== ICP_SYMBOL);
     let networks = $derived.by(() => {
         if (isBtc) {
             return [BTC_SYMBOL, CKBTC_SYMBOL];
         } else if (oneSecEnabled) {
-            return [ICP_SYMBOL, ETHEREUM_NETWORK, ARBITRUM_NETWORK, BASE_NETWORK];
+            return client.oneSecGetNetworks(tokenDetails.symbol);
         } else {
             return [];
         }